# swagger/swagger_output.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "API do BoardTasks",
    "description": "Documentação da API do BoardTasks."
  },
  "servers": [
    {
      "url": "http://localhost:4000/",
      "description": "Servidor localhost."
    },
    {
      "url": "https://dnc-board-tasks-back.vercel.app/",
      "description": "Servidor de produção."
    }
  ],
  "paths": {
    "/livros/criar": {
      "post": {
        "tags": [
          "Livro"
        ],
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "titulo": {
                    "example": "any"
                  },
                  "numeroPaginas": {
                    "example": "any"
                  },
                  "isbn": {
                    "example": "any"
                  },
                  "editora": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/livros/editar/{id}": {
      "put": {
        "tags": [
          "Livro"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "titulo": {
                    "example": "any"
                  },
                  "numeroPaginas": {
                    "example": "any"
                  },
                  "isbn": {
                    "example": "any"
                  },
                  "editora": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/livros/livros": {
      "get": {
        "tags": [
          "Livro"
        ],
        "description": "Endpoint para obter todos livros.",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/livros/obter/{id}": {
      "get": {
        "tags": [
          "Livro"
        ],
        "description": "Endpoint para obter um livro pelo ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Não Encontrado"
          }
        }
      }
    },
    "/livros/deletar/{id}": {
      "delete": {
        "tags": [
          "Livro"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Livro": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Livro"
          },
          "required": {
            "type": "array",
            "example": [
              "titulo",
              "numeroPaginas",
              "isbn",
              "editora"
            ],
            "items": {
              "type": "string"
            }
          },
          "properties": {
            "type": "object",
            "properties": {
              "titulo": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  }
                }
              },
              "numeroPaginas": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "number"
                  }
                }
              },
              "isbn": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  }
                }
              },
              "editora": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  }
                }
              },
              "_id": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  }
                }
              },
              "createdAt": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "format": {
                    "type": "string",
                    "example": "date-time"
                  }
                }
              },
              "updatedAt": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "format": {
                    "type": "string",
                    "example": "date-time"
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}